+++ llmct/utils/__init__.py
@@ -1,5 +1,6 @@
 """工具模块"""
 
 from .logger import get_logger
+from .text_utils import display_width, pad_string, truncate_string
 
-__all__ = ['get_logger']
+__all__ = ['get_logger', 'display_width', 'pad_string', 'truncate_string']